requirements.txt
Fabric3==1.14.post1
python-decouple==3.1